+++ src/android/Callbacks.java
@@ -22,7 +22,6 @@ ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF S
 
 package com.intel.security;
 
-import android.util.Log;
 import android.content.Context;
 import android.provider.Settings;
 import android.content.res.AssetManager;