+++ src/android/SecureData.java
@@ -28,7 +28,6 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import android.util.Base64;
-//import android.util.Log;
 
 public class SecureData {