+++ src/ios/sService.h
@@ -44,7 +44,6 @@ ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF S
 - (void) SecureDataDestroy:(CDVInvokedUrlCommand *)command;
 - (void) SecureStorageRead:(CDVInvokedUrlCommand *)command;
 - (void) SecureStorageWrite:(CDVInvokedUrlCommand *)command;
-- (void) SecureStorageWriteSecureData:(CDVInvokedUrlCommand *)command;
 - (void) SecureStorageDelete:(CDVInvokedUrlCommand *)command;
 - (void) SecureTransportOpen:(CDVInvokedUrlCommand *)command;
 - (void) SecureTransportSetURL:(CDVInvokedUrlCommand *)command;